requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
astroid==2.2.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
dnspython==1.16.0
future==0.17.1
google-auth==1.6.3
gspread==3.1.0
httplib2==0.13.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.0.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pylint==2.3.1
pymongo==3.8.0
python-dotenv==0.10.3
python-telegram-bot==11.1.0
pytz==2019.2
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
wrapt==1.11.2
